fix(ci): restore PR check baseline
