+++ kun/package.json
@@ -98,7 +98,7 @@
     "semver": "^7.8.5",
     "typescript": "5.9.3",
     "typescript-language-server": "5.3.0",
-    "undici": "^7.28.0",
+    "undici": "^7.29.0",
     "yauzl": "^3.4.0",
     "yazl": "^3.3.1",
     "zod": "^4.4.3"
@@ -112,6 +112,9 @@
   },
   "overrides": {
     "@hono/node-server": "2.0.11",
+    "fast-uri": "3.1.5",
+    "hono": "4.13.0",
+    "ip-address": "10.4.0",
     "@connectrpc/connect-node": {
       "undici": "6.28.0"
     }
