+++ kun/src/graph/index.ts
@@ -10,6 +10,8 @@ export * from './graph-artifact-policy.js'
 export * from './graph-worker-context.js'
 export * from './graph-write-coordinator.js'
 export * from './graph-scheduler.js'
+export * from './graph-scheduler-types.js'
+export { parseWorkerResult, validateWorkerResult } from './graph-scheduler-policy.js'
 export * from './graph-supervisor.js'
 export * from './graph-supervision-obligation.js'
 export * from './graph-supervision-view.js'
