+++ package.json
@@ -181,7 +181,10 @@
     "zustand": "^5.0.3"
   },
   "overrides": {
-    "@hono/node-server": "2.0.11"
+    "@hono/node-server": "2.0.11",
+    "fast-uri": "3.1.5",
+    "hono": "4.13.0",
+    "ip-address": "10.4.0"
   },
   "author": "Kun Contributors",
   "license": "PolyForm-Noncommercial-1.0.0",
