+++ src/renderer/src/graph/graph-store.ts
@@ -2,107 +2,25 @@ import { create } from 'zustand'
 import type { RuntimeChildEventPayload } from '../agent/types'
 import { graphRuntimeClient } from './graph-runtime-client'
 import { steerGraphSourceTurn } from './graph-source-turn-steering'
-import { createGraphLeadWakeAction, mergeGraphRunSnapshots } from './graph-supervision-store'
+import type { GraphViewState } from './graph-store-types'
+import {
+  createGraphLeadWakeAction,
+  mergeGraphRunSnapshots
+} from './graph-supervision-store'
 import {
   graphChildRuntimeFromEvent,
   mergeGraphChildDiagnostics,
-  mergeGraphChildRuntime,
-  type GraphChildReturnTarget
+  mergeGraphChildRuntime
 } from './graph-child-runtime'
 import type {
-  GraphAgentEvidence,
   GraphAgentProfile,
-  GraphAgentScore,
-  GraphArtifactPage,
-  GraphChildRuntime,
   GraphEventEnvelope,
-  GraphGovernanceAudit,
-  GraphLearningCandidate,
-  GraphLearningJob,
-  GraphPatchOperation,
   GraphPlanningDraftView,
   GraphPlanningLifecycleEvent,
-  GraphRun,
-  ProjectIdentity
+  GraphRun
 } from './graph-types'
 
 export { selectGraphPlanningCorrectionDraft } from './graph-planning-selection'
-
-type GraphThreadRefreshOptions = {
-  silent?: boolean
-}
-
-type GraphViewState = {
-  threadId: string | null
-  workspace: string
-  runs: GraphRun[]
-  drafts: GraphPlanningDraftView[]
-  childRuns: Record<string, GraphChildRuntime>
-  childReturnTarget: GraphChildReturnTarget | null
-  selectedRunId: string | null
-  selectedNodeId: string | null
-  identity: ProjectIdentity | null
-  profiles: GraphAgentProfile[]
-  evidence: GraphAgentEvidence[]
-  scores: GraphAgentScore[]
-  audit: GraphGovernanceAudit[]
-  candidates: GraphLearningCandidate[]
-  jobs: GraphLearningJob[]
-  exportedProfile: string | null
-  artifactPage: GraphArtifactPage | null
-  artifactContent: string
-  artifactLoading: boolean
-  wakingObligationId: string | null
-  loading: boolean
-  error: string | null
-  refreshThread: (threadId: string | null, options?: GraphThreadRefreshOptions) => Promise<void>
-  refreshProject: (workspace: string) => Promise<void>
-  refreshSelectedRun: () => Promise<void>
-  selectRun: (runId: string | null) => void
-  selectNode: (nodeId: string | null) => void
-  setChildReturnTarget: (target: GraphChildReturnTarget) => void
-  updateChildObserver: (
-    status: GraphChildReturnTarget['observerStatus'],
-    cursor?: number
-  ) => void
-  updateChildSessionStatus: (
-    status: GraphChildReturnTarget['childSessionStatus']
-  ) => void
-  clearChildReturnTarget: () => void
-  receiveChildRuntimeEvent: (event: RuntimeChildEventPayload) => void
-  receiveEvent: (event: GraphEventEnvelope) => void
-  receivePlanningEvent: (event: GraphPlanningLifecycleEvent) => void
-  command: (action: 'start' | 'pause' | 'resume' | 'cleanup') => Promise<void>
-  cancel: () => Promise<void>
-  resumeDraft: (draftId: string) => Promise<void>
-  cancelDraft: (draftId: string) => Promise<void>
-  retryNode: (nodeId: string) => Promise<void>
-  reviewNode: (nodeId: string, outcome: 'pass' | 'fail') => Promise<void>
-  wakeLead: (obligationId?: string) => Promise<void>
-  patch: (operations: GraphPatchOperation[], reason: string) => Promise<void>
-  rebindNode: (nodeId: string, profileId: string) => Promise<void>
-  steer: (text: string, nodeId?: string) => Promise<void>
-  steerSourceTurn: (threadId: string, sourceTurnId: string, text: string) => Promise<boolean>
-  loadArtifact: (artifactId: string) => Promise<void>
-  loadNextArtifactPage: () => Promise<void>
-  clearArtifact: () => void
-  transitionProfile: (
-    profileId: string,
-    lifecycle: GraphAgentProfile['lifecycle']
-  ) => Promise<void>
-  exportProfile: (profileId: string) => Promise<void>
-  importProfile: (portableJson: string) => Promise<void>
-  mergeProfiles: (
-    sourceProfileIds: string[],
-    targetProfileId: string,
-    name: string
-  ) => Promise<void>
-  governCandidate: (
-    candidateId: string,
-    action: 'approve' | 'reject' | 'start_probation' | 'promote' | 'rollback' | 'delete'
-  ) => Promise<void>
-  consolidate: () => Promise<void>
-}
 
 function message(error: unknown): string {
   return error instanceof Error ? error.message : String(error)
